requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-mgmt-resource
openai==0.28
scikit-learn==1.3.0
nltk==3.8.1

